LPS-85514 users-admin-web - refactor helper class package path
