+++ modules/apps/users-admin/users-admin-web/src/main/java/com/liferay/users/admin/web/internal/helper/ContactInformationHelper.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.users.admin.web.internal.portlet.action;
+package com.liferay.users.admin.web.internal.helper;
 
 import javax.portlet.ActionRequest;
